+++ components/planner/index.ts
@@ -11,6 +11,6 @@ export { DailyEventsContainer } from './DailyEventsContainer';
 export { WeeklyEventsDisplay } from './WeeklyEventsDisplay';
 export { MiniDailyTimeline } from './MiniDailyTimeline';
 export { ConflictResolutionModal } from './ConflictResolutionModal';
-export { MergedDailyView } from './MergedDailyView';
+export { MergedDailyView, type MergedDailyViewRenderOptions } from './MergedDailyView';
 export { SchedulingSidebar } from './SchedulingSidebar';